+++ checkoutsys-api/checkoutsys.api/checkoutsys.api/Models/AddProductRequest.cs
@@ -10,6 +10,6 @@
 
         public decimal Price { get; set; }
 
-        public int Stock { get; set; }
+        public long Stock { get; set; }
     }
 }
