+++ checkoutsys-api/checkoutsys.api/checkoutsys.api/Models/PutProductRequest.cs
@@ -8,6 +8,6 @@
 
         public decimal Price { get; set; }
 
-        public int Stock { get; set; }
+        public long Stock { get; set; }
     }
 }
